refactor some functions to base controller class
