--- zmscitizenapi/VERSION
@@ -1,1 +1,0 @@
-2.25.00-muc11-194-g23bcd19c2
